added login and logout pages

+++ projects/page_194_managing_user_sessions/javascript/main.js
@@ -1,3 +1,7 @@
+if (localStorage.getItem("user-token") == null) {
+    window.location.replace(document.location.origin + "/login/");
+}
+
 function renderItems(items, processType, elementId, processFunction) {
     let placeholder = "<div>";
     let itemsMeta = [];
@@ -32,14 +36,18 @@ function apiCall(url, method) {
     xhr.withCredentials = true;
     xhr.addEventListener('readystatechange', function () {
         if (this.readyState === this.DONE) {
-            renderItems(JSON.parse(this.responseText)["pending_items"],
-                "edit", "pendingItems", editItem);
-            renderItems(JSON.parse(this.responseText)["done_items"],
-                "delete", "doneItems", deleteItem);
-            document.getElementById("completeNum").innerHTML =
-                JSON.parse(this.responseText)["done_item_count"];
-            document.getElementById("pendingNum").innerHTML =
-                JSON.parse(this.responseText)["pending_item_count"];
+            if (this.status === 401) {
+                window.location.replace(document.location.origin + "/login/");
+            } else {
+                renderItems(JSON.parse(this.responseText)["pending_items"],
+                    "edit", "pendingItems", editItem);
+                renderItems(JSON.parse(this.responseText)["done_items"],
+                    "delete", "doneItems", deleteItem);
+                document.getElementById("completeNum").innerHTML =
+                    JSON.parse(this.responseText)["done_item_count"];
+                document.getElementById("pendingNum").innerHTML =
+                    JSON.parse(this.responseText)["pending_item_count"];
+            }
         }
     });
     xhr.open(method, url);
